Ensuring that all of the code is being moved around accordingly
